refactor: Add camunda-security-core Module
