+++ authentication/pom.xml
@@ -106,7 +106,7 @@
     </dependency>
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>zeebe-protocol</artifactId>
+      <artifactId>camunda-security-core</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.security</groupId>
